+++ contracts/utils/LinearTokenTimelock.sol
@@ -8,7 +8,7 @@ import "../external/SafeMathCopy.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 contract LinearTokenTimelock is Timed {
-    using SafeMathCopy for uint;
+    using SafeMathCopy for uint256;
 
     // ERC20 basic token contract being held
     IERC20 public lockedToken;
@@ -18,17 +18,24 @@ contract LinearTokenTimelock is Timed {
 
     address public pendingBeneficiary;
 
-    uint public initialBalance;
+    uint256 public initialBalance;
 
-    uint internal lastBalance;
+    uint256 internal lastBalance;
 
-    event Release(address indexed _beneficiary, uint _amount);
+    event Release(address indexed _beneficiary, uint256 _amount);
     event BeneficiaryUpdate(address indexed _beneficiary);
     event PendingBeneficiaryUpdate(address indexed _pendingBeneficiary);
 
-    constructor (address _beneficiary, uint _duration, address _lockedToken) public Timed(_duration) {
+    constructor(
+        address _beneficiary,
+        uint256 _duration,
+        address _lockedToken
+    ) public Timed(_duration) {
         require(_duration != 0, "LinearTokenTimelock: duration is 0");
-        require(_beneficiary != address(0), "LinearTokenTimelock: Beneficiary must not be 0 address");
+        require(
+            _beneficiary != address(0),
+            "LinearTokenTimelock: Beneficiary must not be 0 address"
+        );
 
         beneficiary = _beneficiary;
         _initTimed();
@@ -39,7 +46,7 @@ contract LinearTokenTimelock is Timed {
     // Prevents incoming LP tokens from messing up calculations
     modifier balanceCheck() {
         if (totalToken() > lastBalance) {
-            uint delta = totalToken().sub(lastBalance);
+            uint256 delta = totalToken().sub(lastBalance);
             initialBalance = initialBalance.add(delta);
         }
         _;
@@ -47,36 +54,42 @@ contract LinearTokenTimelock is Timed {
     }
 
     modifier onlyBeneficiary() {
-        require(msg.sender == beneficiary, "LinearTokenTimelock: Caller is not a beneficiary");
+        require(
+            msg.sender == beneficiary,
+            "LinearTokenTimelock: Caller is not a beneficiary"
+        );
         _;
     }
 
     function release() external onlyBeneficiary balanceCheck {
-        uint amount = availableForRelease();
+        uint256 amount = availableForRelease();
         require(amount != 0, "LinearTokenTimelock: no tokens to release");
 
         lockedToken.transfer(beneficiary, amount);
         emit Release(beneficiary, amount);
     }
 
-    function totalToken() public view virtual returns(uint) {
+    function totalToken() public view virtual returns (uint256) {
         return lockedToken.balanceOf(address(this));
     }
 
-    function alreadyReleasedAmount() public view returns(uint) {
+    function alreadyReleasedAmount() public view returns (uint256) {
         return initialBalance.sub(totalToken());
     }
 
-    function availableForRelease() public view returns(uint) {
-        uint elapsed = timeSinceStart();
-        uint _duration = duration;
+    function availableForRelease() public view returns (uint256) {
+        uint256 elapsed = timeSinceStart();
+        uint256 _duration = duration;
 
-        uint totalAvailable = initialBalance.mul(elapsed) / _duration;
-        uint netAvailable = totalAvailable.sub(alreadyReleasedAmount());
+        uint256 totalAvailable = initialBalance.mul(elapsed) / _duration;
+        uint256 netAvailable = totalAvailable.sub(alreadyReleasedAmount());
         return netAvailable;
     }
 
-    function setPendingBeneficiary(address _pendingBeneficiary) public onlyBeneficiary {
+    function setPendingBeneficiary(address _pendingBeneficiary)
+        public
+        onlyBeneficiary
+    {
         pendingBeneficiary = _pendingBeneficiary;
         emit PendingBeneficiaryUpdate(_pendingBeneficiary);
     }
@@ -86,7 +99,10 @@ contract LinearTokenTimelock is Timed {
     }
 
     function _setBeneficiary(address newBeneficiary) internal {
-        require(newBeneficiary == pendingBeneficiary, "LinearTokenTimelock: Caller is not pending beneficiary");
+        require(
+            newBeneficiary == pendingBeneficiary,
+            "LinearTokenTimelock: Caller is not pending beneficiary"
+        );
         beneficiary = newBeneficiary;
         emit BeneficiaryUpdate(newBeneficiary);
         pendingBeneficiary = address(0);
