+++ contracts/utils/Timed.sol
@@ -7,15 +7,15 @@ import "../external/SafeMathCopy.sol";
 /// @title an abstract contract for timed events
 /// @author Fei Protocol
 abstract contract Timed {
-    using SafeCast for uint;
+    using SafeCast for uint256;
 
     /// @notice the start timestamp of the timed period
-    uint public startTime;
+    uint256 public startTime;
 
     /// @notice the duration of the timed period
-	uint public duration;
+    uint256 public duration;
 
-    constructor(uint _duration) public {
+    constructor(uint256 _duration) public {
         duration = _duration;
     }
 
@@ -26,18 +26,18 @@ abstract contract Timed {
 
     /// @notice number of seconds remaining until time is up
     /// @return remaining
-    function remainingTime() public view returns (uint) {
+    function remainingTime() public view returns (uint256) {
         return SafeMathCopy.sub(duration, timeSinceStart());
     }
 
     /// @notice number of seconds since contract was initialized
     /// @return timestamp
     /// @dev will be less than or equal to duration
-    function timeSinceStart() public view returns (uint) {
-		uint _duration = duration;
-		// solhint-disable-next-line not-rely-on-time
-		uint timePassed = SafeMathCopy.sub(block.timestamp, startTime);
-		return timePassed > _duration ? _duration : timePassed;
+    function timeSinceStart() public view returns (uint256) {
+        uint256 _duration = duration;
+        // solhint-disable-next-line not-rely-on-time
+        uint256 timePassed = SafeMathCopy.sub(block.timestamp, startTime);
+        return timePassed > _duration ? _duration : timePassed;
     }
 
     function _initTimed() internal {
